+++ es6/index.js
@@ -4,15 +4,15 @@ module.exports = {
 
   env: {
     es6: true,
-    node: true,
+    node: true
   },
 
   ecmaFeatures: {
-    modules: true,
+    modules: true
   },
 
   rules: Object.assign({},
     require('../rules/ecmascript-6'),
     {'no-param-reassign': 0} // because of default params
-  ),
+  )
 };