+++ rules/best-practices.js
@@ -117,5 +117,5 @@ module.exports = {
   // Require immediate function invocation to be wrapped in parentheses
   'wrap-iife': [1, 'inside'],
   // Require or disallow Yoda conditions
-  'yoda': [1, 'never'],
+  'yoda': [1, 'never']
 };