+++ rules/node.js
@@ -18,5 +18,5 @@ module.exports = {
   // Restrict usage of specified node modules
   'no-restricted-modules': 0,
   // Disallow use of synchronous methods
-  'no-sync': 1,
+  'no-sync': 1
 };