+++ rules/stylistic-issues.js
@@ -108,5 +108,5 @@ module.exports = {
   // Require or disallow a space immediately following the // or /* in a comment
   'spaced-comment': [1, 'always'],
   // Require regex literals to be wrapped in parentheses
-  'wrap-regex': 0,
+  'wrap-regex': 0
 };