+++ CHANGELOG.md
@@ -13,6 +13,9 @@ All changes to this project are documented in this file. This project adheres to
 - Update `.eslintrc.js` to extend `ryansobol/es5`
 - Extend `ryansobol/core` for ES5, ES6, and React rules
 
+#### Possible Errors
+- Change `comma-dangle` rule to `'never'`
+
 ### Removed
 - Replace `merge` dependency with `Object.assign()` function
 - Remove `eslint-plugin-shopify` rules and dependencies